# Speed up IV writes