--- a/shape.c
+++ b/shape.c
@@ -312,18 +312,6 @@ rb_shape_id_offset(void)
     return 8 - rb_shape_id_num_bits() / 8;
 }
 
-uint8_t
-rb_shape_flag_shift(void)
-{
-    return SHAPE_FLAG_SHIFT;
-}
-
-uint64_t
-rb_shape_flag_mask(void)
-{
-    return SHAPE_FLAG_MASK;
-}
-
 rb_shape_t *
 rb_shape_rebuild_shape(rb_shape_t * initial_shape, rb_shape_t * dest_shape)
 {