--- a/yjit/bindgen/src/main.rs
+++ b/yjit/bindgen/src/main.rs
@@ -95,6 +95,7 @@ fn main() {
 
         // From ruby/internal/intern/object.h
         .allowlist_function("rb_obj_is_kind_of")
+        .allowlist_function("rb_obj_frozen_p")
 
         // From ruby/internal/encoding/encoding.h
         .allowlist_type("ruby_encoding_consts")