@tenderlove 2022-11-18
I wanted to measure the impact of this patch when writing non-immediate instance variables, so I made [a benchmark to measure it](https://github.com/Shopify/yjit-bench/pull/135).

We get a nice speed up there, though not as good as the immediate case:

```
before: ruby 3.2.0dev (2022-11-18T20:04:10Z master 9e067df76b) +YJIT [x86_64-linux]
after: ruby 3.2.0dev (2022-11-18T21:13:32Z iv-write 6f7f05fb68) +YJIT [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  ------------  -------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
setivar_object  50.4         0.2         32.0        0.2         1.57          1.07         
--------------  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```