@tenderlove 2022-12-04
Thanks for merging this.  Wow!

<img width="1051" alt="YJIT Benchmarks | Benchmarks and continuous checks for YJIT performance  2022-12-03 20-50-45" src="https://user-images.githubusercontent.com/3124/205475096-6ff6e1c1-1fc4-4c5f-b267-c4b0163ffbf4.png">